+++ nym-vpn-core/crates/nym-vpn-account-controller/src/state_machine/logged_out_state.rs
@@ -83,6 +83,9 @@ impl<C: ConnectivityMonitor> AccountControllerStateHandler<C> for LoggedOutState
                         shared_state.firewall_active = true;
                         return_sender.send(Ok(()));
                     },
+                    AccountCommand::SetRefreshMode(mode) => {
+                        shared_state.refresh_mode = mode;
+                    },
 
                     AccountCommand::Common(common_command) => {
                         match common_command {