+++ nym-vpn-core/crates/nym-vpn-account-controller/src/state_machine/upgrade_mode_state.rs
@@ -186,6 +186,9 @@ impl UpgradeModeState {
                 shared_state.firewall_active = true;
                 return_sender.send(Ok(()));
             }
+            AccountCommand::SetRefreshMode(mode) => {
+                shared_state.refresh_mode = mode;
+            }
             AccountCommand::Common(common_command) => {
                 common_handler::handle_common_command(common_command, shared_state).await
             }